实现下推过Reorder
